fix 32-bit types to long not long long

+++ util/mspop.c
@@ -16,10 +16,10 @@ https://creativecommons.org/publicdomain/zero/1.0/legalcode
 
 typedef signed char int8_t;
 typedef signed short int16_t;
-typedef signed long long int32_t;
+typedef signed long int32_t;
 typedef unsigned char uint8_t;
 typedef unsigned short uint16_t;
-typedef unsigned long long uint32_t;
+typedef unsigned long uint32_t;
 
 #if defined(__32BIT__)
 #define inpb(p) PosInp(p)
